vulkan: Always add VK_MVK_moltenvk to device extension list if supported.
Vulkan loader filters device extension list to only allow extensions
defined in the Vulkan registry xml file, and VK_MVK_moltenvk is not
included in Vulkan loader.

This change will always add VK_MVK_moltenvk to the device extension support
list as long as it is supported on host (it may be loaded as secondary
library functions, skipping Vulkan loaders and layers). In that case AEMU
guests still need to know that the extension is supported, so we add the
extension at the end of the |properties| list.

Manual cherry-pick from aosp/1495918

Change-Id: Icdf6bd7c695935065b241994d945e17cad299d6f