darwin: Use MTLTexture to share ColorBuffer between Vulkan instances
Previously we have used IOSurface to share Vulkan-backed ColorBuffers.
This has a few drawbacks:

* After creating IOSurfaces, the VkImage will be mapped to MTLTexture
  using the IOSurface, instead of using MTLBuffers from the
    VkDeviceMemory; so host memory mapping doesn't work well.

* Host mapping of IOSurface requires lock and unlock operations
  which don't work well with guest mappings and can be a not
    particularly cheap operation.

MoltenVK also has an API to export the internal MTLTexture of a
VkImage (vkGetMTLSurfaceMVK) and import it into another image
(vkSetMTLSurfaceMVK). This makes it possible to share host-visible
images, and sharing MTL surfaces will be supported as a standard
Vulkan extension in the future (see
https://github.com/KhronosGroup/MoltenVK/issues/830 ).

Manual cherry-pick from aosp/1463002

TEST=spinning-square-rs using
     MVK_CONFIG_FORCE_LOW_POWER_GPU=1 fx emu -N --host-gpu -e <emu-dir>/objs

Change-Id: I291e1ecb14cdf750376af10baa19049d9d0ff56e