vulkan-cereal: Remove VkExternalMemoryProperties from auto-transformed types.
Instead of using auto-generated transform functions, we implement
the fromhost transformation function of VkExternalMemoryProperties
in ResourceTracker separately (in change aosp/1574707) using a
slightly different logic.

However, though VkExternalMemoryProperties is not using default
transform functions, it should be still listed as a transformed
type so that we can generate transformation code in VkEncoder.

Bug: fuchsia:60064

Change-Id: I28078d53996d59ffe08e7dd4f3b50460e802a43b