[1/n] batched descriptor update: codegen
Bug: 177241396

Add new commands

vkQueueCommitDescriptorSetUpdates
vkCollectDescriptorPoolIds

The first is used to commit the descriptor set updates to the host,
while the second is so the guest can allocate a bunch of virtual
descriptor set host handles without talking to the host for each one.

This also requires wrapping of:

vkCmdBindDescriptorSets
vKDestroyDescriptorSetLayout

in the guest. In the first case, we need to track the relevant
descriptor sets to commit for a given submission. In the second case,
since we've modified descriptor set lifetimes on the host driver we need
to ensure the layouts aren't destroyed until the sets are destroyed.

Change-Id: I10306951533071150e88389e6fb604eb0ad5ab62