Fix YUV transfer size calculation
Previously, this seemed to treat all YUV as tri-planar. For a
1440x1080 NV12 buffer, the previous calculation would compute a
uvStride of 736 and then compute the total uvSize as
2 * (uvStride=736 * uvHeight=540) which would effectively double
any extra padding due to alignment. NV12 buffers have an
interleaved UV plane so the align should happen on the combined
width which in this example would result in a width and stride of
1440.

Bug: b/180405004
Test: cts -m CtsCameraTestCases -t \
android.hardware.camera2.cts.AllocationTest#testAllocationFromCameraFlexibleYuv[1]
Change-Id: I145b70f14e79be2409f5a495333443af68a99917

Change-Id: Iae6d1a24b4ed6968ac68779896832822464ba6f9