stream-servers: explicitly zero-initialize variables
Clang's static analyzer complains that `createSharedTrivialContext`
might not initialize these properly, which seems to be correct, since
this function has an early exit if we fail to get a proper config
object.

We have a compiler option that zero-initializes all stack variables
anyway, so this is no functional change. Still, it seems best to init
these variables to make our tooling happy & minimize reliance on this
option going forward.

Bug: 206470603
Test: TreeHugger

Change-Id: I56c0e73c8d91105cb6d59b8e2f9cb98739905961