Vulkan: Redesign how QueueSignalReleaseImage synchronization works
Introduce the VkQsriTimeline class with 2 interfaces, registerCallback
and signal. When calling registerCallback, the current
requestPresentCount will be associated with the callback, the
requestPresentCount will be increased by 1, and call all callbacks whose
associated requestPresentCount is smaller or equal to the current
presentCount. When calling signal, presentCount will be increased by 1,
and callbacks will be scanned and called in the same way. Every anb will
has its own VkQsriTimeline.

The new design will call callbacks right after the callback is
registered or the present is signaled on the same thread to prevent from
dispatching 2 tasks with dependency in the wrong order(first callback
waiting for present signal, then signal the present) to the same
SyncThread worker thread.

Test: compile and run.
Change-Id: I6958e29373b12ddd0cb188755cae9cc0a38f3bc9