Updates YUVConverter bi-planar and tri-planar handling
... to always have three texture variables Y, U, and V with U and
V referring to the same texture name in the bi-planar case and to
have a single shader creation function which generates different
shaders based on format properties. This helps avoid
`switch`-s in various functions and helps avoiding additional
blow up when adding for new formats later.

This is part of a chain of changes to refactor YUVConverter
to support YUV P010 buffers.

Bug: b/191084459
Test: cvd start --gpu_mode=gfxstream
Change-Id: Ib79dfb560cf574f238c8832e21a1887d790b5d0b