Minor fix to OpenGL test:
This was a parameterized test using a pointer as the parameter type. As a result, the name of the test would be different from one machine to another because CTest would use the pointer value as part of the test name (for example: `GLES2SnapshotRasterization/SnapshotGlPolygonOffsetTest.SetPolygonOffset/00007FF674529148`), which in turn didn't work with our mechanism to skip certain tests on Kokoro.

Test: compile, run the test
Change-Id: I98e5f4c11bdc87c1c65fe23f8065bc932b5b2d8a