Reset external fences by adding them to the pool and replacing them.
In order to prevent the potential situation in which an external VkFence is
reset before waitForFence is called on it, we don't reset external fences.
Instead, we add it to the external fence pool and opaquely swap out the
underlying fence for one that we know is ready for use.

Test: Built and ran B*, ran tests.
Bug: 232444664
Change-Id: Iab1d3f03a4b13b158a526d71478f6e0ea0f29fc4