Perform ASTC texture decompression on the CPU.
We currently have a compute shader to decompress ASTC textures on the
GPU, but it's not well-maintained and crashes in certain cases. On
larger textures, for example, it might take too long to run and cause a
device lost error.

As a result, we're experimenting with CPU texture decompression, using
ARM's ASTC encoder / decoder library. Speed is around 20M texels/s, and
can be further improved by using multithreaded decompression.

There are, however, a couple of issues with this approach:

 1. it uses more memory since we're passing uncompressed textures to the
 GPU. We can mitigate that by re-compressing to a format supported by
 the GPU, at the cost of more processing time.

 2. it may not always work. In certain cases, the texture data may not
 be available by the time we try to decompress it.

This is still experimental and we're currently gating this feature
behind a flag.

BUG=238656926

Change-Id: Ibec1370c08639c42efc9b0b2909b935c8a7257a6