Introduce gfxstream_gl_server library
... , to parallel the `vulkan` subdirectory and
`gfxstream_vulkan_server` library, and move `FbConfig`
into the new subdirectory as `EmulatedEglConfig`.

This change is part of a series of changes to make it
possible to disable GL on the host when running with
guest ANGLE. For this effort, we want to effectively wrap
most of the existing logic into a "GlEmulation" which is
not initialized when running with guest ANGLE.

Also, replace some raw pointer usage with std::vector.

Bug: b/233939967
Test: android build
Test: cmake build
Test: cvd start --gpu_mode=gfxstream
Test: gfxstream unit tests
Change-Id: Ifa074c285ce710914218ae29db71e53280e97d67