Add DisplayGl and DisplaySurfaceGl
... and update handling of the created EGL surfaces and contexts.

Moves the existing Display interface to DisplaySurfaceUser as
CompositorGl needs to be bound to either the "fake window" or the
"real window" surface.

Removes TextureDraw from ContextHelper as there is only ever a
single TextureDraw instance created using the global EGL context
which is shared with every other EGL context.

Moves the existing GL ops out of PostWorker and into DisplayGl.
Upcoming changes hope to consolidate DisplayGl and DisplayVk
interface so that PostWorker does not have any knowledge of the
API specific displays.

Bug: b/233939967
Test: android build
Test: cmake build
Test: cvd start --gpu_mode=gfxstream
Test: gfxstream unit tests
Change-Id: I0d5ad35978df5b89b510e76ac7a16e59c6aef57d