Fix YUV transfer size for non-aligned widths
The virtio gpu transfer size calculation was using the widths
instead of the strides to compute the total size of transfer
needed. This would result in missing data when a buffer had
extra padding for alignment. One example of this is watching
a Youtube short video, which is "portrait" sized, and uses
a 368x640 YV12 buffer which is padded to 384x640.

Bug: b/254425230
Test: cvd start --gpu_mode=gfxstream
Test: Open landscape Youtube video in web browser
Test: Open portrait Youtube short video in web browser
Test: Open camera app
Test: cts -m CtsNativeHardwareTestCases
Change-Id: If23fa6ac47b469bb4a771c4a710e47622236de83