Make FrameBuffer, not ReadbackWorker, invoke post callback
Upcoming refactoring aims to hide GL/EGL internals from the
generic classes to support running on a host without GL and
only vulkan. For this, ReadbackWorker will become an interface
and the various backends will implement ReadbackWorkerGl/
ReadbackWorkerVk. In order to avoid a circular dependency
between FrameBuffer <-> ReadbackWorker, the invocation of the
post callback should be moved to FrameBuffer.

Bug: b/233939967
Test: android build
Test: cmake build
Change-Id: I33e141e25cd29a72f9cff09f85213cf54a57cfae